Include read exception tests in test_read_flux
