+++ tests/test_read_flux.py
@@ -2,15 +2,20 @@
 Tests reading thermal flux and calculating thermal conductivity
 """
 import os
+import pytest
 import yaml
 import numpy as np
-from teemof.read import read_thermal_flux, calculate_k, estimate_k, average_k
+from teemof.read import read_thermal_flux, calculate_k, estimate_k, average_k, get_flux_directions
+from teemof.read import FluxFileNotFoundError, TimestepsMismatchError
 
 
 flux_file = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'thermal-flux.dat')
 flux_ref_file = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'flux.yaml')
 time_ref_file = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'time.yaml')
 k_ref_file = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'thermal-conductivity.yaml')
+trial_dir = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'ideal-mof-trial')
+k_parameters = dict(kb=0.001987, conv=69443.84, dt=5, volume=80 * 80 * 80,
+                    temp=300, prefix='J0Jt_t', isotropic=False, average=True)
 
 
 def test_thermal_flux_calculation():
@@ -40,7 +45,23 @@ def test_thermal_conductivity_average():
 
 def test_thermal_conductivity_estimation():
     flux, time = read_thermal_flux(flux_file)
-    k_parameters = dict(kb=0.001987, conv=69443.84, dt=5, volume=80 * 80 * 80, temp=300)
     J = calculate_k(flux, k_par=k_parameters)
     k = estimate_k(J, time, t0=5, t1=10)
     assert np.isclose(k, 0.8778570946468635)
+
+
+def test_get_flux_directions_exception():
+    """Tests whether thermal flux files note found exception is raised correctly"""
+    k_parameters['prefix'] = 'wrong-name'
+    with pytest.raises(FluxFileNotFoundError):
+        get_flux_directions(os.path.join(trial_dir, 'Run1'), k_par=k_parameters)
+
+
+def test_average_k_exception():
+    """Tests whether trying to average thermal conductivity results with different
+    number of timesteps raises exception correctly"""
+    with open(k_ref_file, 'r') as kref:
+        k_ref = yaml.load(kref)
+    k_runs = [k_ref, k_ref[:10], k_ref[5:]]
+    with pytest.raises(TimestepsMismatchError):
+        average_k(k_runs)
